refac: store time as str & convert to min to calc
Previously times were stored as local ms. This was overly complicated
and resulted in issues when calculating times due to rounding errors. By
storing everything as a time string, we can avoid extra conversions and
since everything is relative to the minute there is no need to maintain
ms resolution.